requirements.txt
torch>=2.1.0
torchaudio>=2.1.0
librosa>=0.10.0
numpy>=1.24.0
gradio>=3.40.0
pyyaml>=6.0
tqdm>=4.65.0
